+++ src/components/goal/HabitStats.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
 import { computed, defineProps } from 'vue';
+import { type Day } from 'date-fns';
 
-import { Goal } from 'src/lib/api/goal.ts';
+import type { HabitGoal } from 'server/lib/models/goal/types';
 import { Tally } from 'src/lib/api/tally.ts';
 
 import { analyzeStreaksForHabit } from 'server/lib/models/goal/helpers';
@@ -11,10 +12,13 @@ import { commaify } from 'src/lib/number.ts';
 import { type HabitGoalParameters } from 'server/lib/models/goal/types';
 import { GOAL_CADENCE_UNIT_INFO } from 'server/lib/models/goal/consts';
 
-const props = defineProps<{
-  goal: Goal;
+const props = withDefaults(defineProps<{
+  goal: HabitGoal;
   tallies: Tally[];
-}>();
+  weekStartsOn?: Day;
+}>(), {
+  weekStartsOn: 0, // Sunday
+});
 
 const habitStats = computed(() => {
   const parameters = props.goal.parameters as HabitGoalParameters;
@@ -24,6 +28,7 @@ const habitStats = computed(() => {
     parameters.threshold,
     props.goal.startDate,
     props.goal.endDate,
+    props.weekStartsOn,
   );
 
   return stats;